clean up module organization
