+++ src/groundhog_hpc/__init__.py
@@ -1,4 +1,3 @@
-from groundhog_hpc.decorators.function import _function as function
-from groundhog_hpc.decorators.harness import _harness as harness
+from groundhog_hpc.decorators import function, harness
 
 __all__ = ["function", "harness"]
